rootdir: enable msm_irqbalance
I did some careful profiling with Qualcomm Trepn and found a significant
measurable improvement in CPU load distribution. Distributing IRQs is
useful for reducing latency and cache misses. We have 8 cores here, so
we need to spread the work and not have it all on CPU 0.

I compared average load on CPUs 0, 3, 4, and 7 with Trepn.

Without msm_irqbalance, average loads on the selected CPUs were:
22%	0%	18%	1%

With msm_irqbalance, this improved to:
11%	2%	13%	8%