gps: use prebuilt HAL for now
The source built one is causing reboots when mobile data is connected,
and the prebuilt one doesn't. The prebuilt one is linked with
libmdmdetect and libperipheral_client and probably does something
relevant with them that the source built CM one doesn't. As with
audio, we should try to figure out the culprit and fix it in the
source built one, but for now I don't like crashes, so let's just
use the prebuilt one.